+++ src/components/button-toggle-group/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./button-toggle-group.component";
+export type { ButtonToggleGroupProps } from "./button-toggle-group.component";