+++ src/components/menu/menu-item/menu-item.d.ts
@@ -29,6 +29,10 @@ export interface MenuItemBaseProps extends LayoutProps, FlexboxProps {
   onSubmenuOpen?: () => void;
   /** Callback triggered when submenu closes. Only valid with submenu prop */
   onSubmenuClose?: () => void;
+  /** @ignore @private
+  private prop, used inside ScrollableBlock to ensure the MenuItem's color variant overrides the CSS
+  for other MenuItems inside the block */
+  overrideColor?: boolean;
 }
 
 export interface MenuWithChildren extends MenuItemBaseProps {