+++ src/components/menu/scrollable-block/scrollable-block.component.js
@@ -2,6 +2,7 @@ import React, { useContext } from "react";
 import PropTypes from "prop-types";
 
 import MenuContext from "../menu.context";
+import MenuItem from "../menu-item";
 import SubmenuContext from "../__internal__/submenu/submenu.context";
 import StyledScrollableBlock from "./scrollable-block.style";
 import Box from "../../box";
@@ -10,6 +11,8 @@ const ScrollableBlock = ({
   children,
   height,
   variant = "default",
+  parent,
+  parentVariant,
   ...rest
 }) => {
   const menuContext = useContext(MenuContext);
@@ -28,15 +31,20 @@ const ScrollableBlock = ({
       data-component="submenu-scrollable-block"
       menuType={menuContext.menuType}
       variant={variant}
-      role="presentation"
       {...rest}
     >
+      {parent && (
+        <MenuItem overrideColor variant={parentVariant} as="div" href="#">
+          {parent}
+        </MenuItem>
+      )}
       <Box
         overflowY="scroll"
         scrollVariant={scrollVariants[menuContext.menuType]}
         height={height}
         p={0}
-        role="presentation"
+        as="ul"
+        role="list"
       >
         {React.Children.map(children, (child, index) => {
           let isFocused = false;
@@ -69,6 +77,12 @@ ScrollableBlock.propTypes = {
   height: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   /** set the colour variant for a menuType */
   variant: PropTypes.oneOf(["default", "alternate"]),
+  /** the element, if any, displayed at the top of the block to be its semantic "parent",
+   * but not part of the scrollable section
+   */
+  parent: PropTypes.node,
+  /** the colour variant for the parent element, if different from the variant of the block */
+  parentVariant: PropTypes.oneOf(["default", "alternate"]),
 };
 
 export default ScrollableBlock;